build: improve doc-generation script
